+++ lib/index.mjs
@@ -514,7 +514,7 @@ function generate(tokens) {
   function flushLine(tokens) {
     linesHtml.push(createLine(
       tokens.map(([type, value]) => (
-        `<span style="color: var(--sh-${types[type]})">${encode(value)}</span>`
+        `<span class="sh__token--${types[type]}" style="color: var(--sh-${types[type]})">${encode(value)}</span>`
       ))
       .join('')
     ))
